requirements.txt
tabulate
pandas
fpdf2